LPS-116569 Move Ratings.js
